phase_4_2_prod_wiring/requirements.txt
psycopg[binary]>=3.1
pgvector>=0.3.0
boto3>=1.34
sentence-transformers>=2.7
pyyaml>=6.0
pytest>=8.0
moto[s3]>=5.0
# openai>=1.12  # only needed if switching back to the openai provider
